[ANDROSDK-624] Rename sharedTest package
